Update whenever gem roles for verify cron
By default, the whenever gem only deploys to servers with the :db role. Based on the config/environments/[environment].rb, the cron is only being pushed to the first server in the list. To rectify the problem, I'm updating the cron to push to all servers that get the :app role